Refactor eth_feeHistory spec
